Add Optuna training script and refactor LSTM utilities
